make process results a utility
