+++ redisvl/utils/utils.py
@@ -1,5 +1,9 @@
 import re
-from typing import Any, List, Optional, Pattern
+from typing import TYPE_CHECKING, Any, Dict, List, Optional, Pattern
+
+if TYPE_CHECKING:
+    from redis.commands.search.result import Result
+    from redis.commands.search.document import Document
 
 import numpy as np
 
@@ -78,4 +82,17 @@ def check_redis_modules_exist(client) -> None:
 
 
 def array_to_buffer(array: List[float], dtype: Any = np.float32) -> bytes:
+    """Convert a list of floats into a numpy byte string."""
     return np.array(array).astype(dtype).tobytes()
+
+
+def process_results(results: List["Result"]) -> List[Dict[str, Any]]:
+    """Convert a list of search Result objects into a list of document dicts"""
+
+    def _process(doc: "Document") -> dict:
+        d = doc.__dict__
+        if "payload" in d:
+            del d["payload"]
+        return d
+
+    return [_process(doc) for doc in results.docs]
